Move !about to info cog.
